+++ router/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const userRouter = require('./user.js');
+
+const router = express.Router();
+
+router.use('/', userRouter);
+
+module.exports = router;
